add list menu

+++ src/List/List.ts
@@ -12,15 +12,18 @@ export interface List {
 export const createList = async (list: List) => {
     const listRef = firestore.collection("lists");
     try {
-        listRef.add({
-            ...list,
-        });
-        list.users.forEach(async (user) => {
-            const userRef = firestore.collection("users").doc(user);
-            await userRef.update({
-                toDos: fieldValue.arrayUnion(listRef.id),
+        listRef
+            .add({
+                ...list,
+            })
+            .then((newList) => {
+                list.users.forEach(async (user) => {
+                    const userRef = firestore.collection("users").doc(user);
+                    await userRef.update({
+                        toDos: fieldValue.arrayUnion(newList.id),
+                    });
+                });
             });
-        });
         console.log("list created.");
     } catch (error) {
         console.error(error.message);
@@ -48,9 +51,16 @@ export const updateList = async (list: List) => {
     }
 };
 
-export const deleteList = async (listId: string) => {
-    const listRef = firestore.collection("lists").doc(listId);
+export const deleteList = async (list: List) => {
+    if (!list) return;
+    const listRef = firestore.collection("lists").doc(list.id);
     try {
+        list.users.forEach((user) => {
+            const userRef = firestore.collection("users").doc(user);
+            userRef.update({
+                toDos: fieldValue.arrayRemove(list?.id),
+            });
+        });
         await listRef.delete();
         console.log("list delete done.");
     } catch (error) {
